+++ tests/Command/CommandTest.php
@@ -4,9 +4,9 @@ namespace Test\OctoLab\Cilex\Command;
 
 use Cilex\Application;
 use OctoLab\Cilex\Command\Command;
-use OctoLab\Cilex\Provider\ConfigServiceProvider;
-use OctoLab\Cilex\Provider\DoctrineServiceProvider;
-use OctoLab\Cilex\Provider\MonologServiceProvider;
+use OctoLab\Cilex\ServiceProvider\ConfigServiceProvider;
+use OctoLab\Cilex\ServiceProvider\DoctrineServiceProvider;
+use OctoLab\Cilex\ServiceProvider\MonologServiceProvider;
 use Test\OctoLab\Cilex\TestCase;
 use Symfony\Component\Console\Output\NullOutput;